All code finished except summaries

+++ gp/a2c/models/base_model.py
@@ -17,7 +17,7 @@ class BaseModel:
         :return:
         """
         with tf.variable_scope('global_step'):
-            self.global_step_tensor = tf.Variable(0, trainable=False, name='global_step')
+            self.global_step_tensor = tf.Variable(0, trainable=False, name='global_step', dtype='int32')
             self.global_step_input = tf.placeholder('int32', None, name='global_step_input')
             self.global_step_assign_op = self.global_step_tensor.assign(self.global_step_input)
